update error handler, controller, routes
